+++ python/sglang/srt/managers/scheduler.py
@@ -24,6 +24,26 @@ from concurrent import futures
 from dataclasses import dataclass
 from typing import Dict, List, Optional, Tuple, Union, Callable
 
+import logging
+import os
+import signal
+from typing import Optional
+
+import psutil
+import setproctitle
+import zmq
+
+from sglang.srt.managers.scheduler.communicator import SchedulerCommunicator
+from sglang.srt.managers.scheduler.core import SchedulerCore
+from sglang.srt.server_args import PortArgs, ServerArgs
+from sglang.srt.utils import (
+    configure_logger,
+    get_bool_env_var,
+    set_gpu_proc_affinity,
+    suppress_other_loggers, get_zmq_socket, broadcast_pyobj,
+)
+from sglang.utils import get_exception_traceback, TypeBasedDispatcher
+
 import psutil
 import torch
 from sglang.global_config import global_config
@@ -79,17 +99,17 @@ logger = logging.getLogger(__name__)
 test_retract = get_bool_env_var("SGLANG_TEST_RETRACT")
 
 
-# TODO improve naming...
-class SchedulerCore:
+class Scheduler:
     """A scheduler that manages a tensor parallel GPU worker."""
 
     def __init__(
-        self,
-        server_args: ServerArgs,
-        nccl_port: int,
-        gpu_id: int,
-        tp_rank: int,
-        dp_rank: Optional[int],
+            self,
+            server_args: ServerArgs,
+            port_args: PortArgs,
+            nccl_port: int,
+            gpu_id: int,
+            tp_rank: int,
+            dp_rank: Optional[int],
     ):
         # Parse args
         self.server_args = server_args
@@ -214,8 +234,8 @@ class SchedulerCore:
         self.req_to_token_pool, self.token_to_kv_pool = self.tp_worker.get_memory_pool()
 
         if (
-            server_args.chunked_prefill_size is not None
-            and server_args.disable_radix_cache
+                server_args.chunked_prefill_size is not None
+                and server_args.disable_radix_cache
         ):
             self.tree_cache = ChunkCache(
                 req_to_token_pool=self.req_to_token_pool,
@@ -229,6 +249,31 @@ class SchedulerCore:
             )
         self.tree_cache_metrics = {"total": 0, "hit": 0}
         self.policy = SchedulePolicy(self.schedule_policy, self.tree_cache)
+
+        context = zmq.Context(2)
+
+        if self.tp_rank == 0 or self.server_args.enable_dp_attention:
+            self._recv_from_tokenizer = get_zmq_socket(
+                context, zmq.PULL, port_args.scheduler_input_ipc_name
+            )
+            self._send_to_tokenizer = get_zmq_socket(
+                context, zmq.PUSH, port_args.tokenizer_ipc_name
+            )
+
+            if self.server_args.skip_tokenizer_init:
+                # Directly send to the TokenizerManager
+                self._send_to_detokenizer = get_zmq_socket(
+                    context, zmq.PUSH, port_args.tokenizer_ipc_name
+                )
+            else:
+                # Send to the DetokenizerManager
+                self._send_to_detokenizer = get_zmq_socket(
+                    context, zmq.PUSH, port_args.detokenizer_ipc_name
+                )
+        else:
+            self._recv_from_tokenizer = None
+            self._send_to_tokenizer = SimpleNamespace(send_pyobj=lambda x: None)
+            self._send_to_detokenizer = SimpleNamespace(send_pyobj=lambda x: None)
 
         # Init running status
         self.waiting_queue: List[Req] = []
@@ -254,7 +299,7 @@ class SchedulerCore:
             self.chunked_prefill_size = None
         self.being_chunked_req = None
         self.is_mixed_chunk = (
-            self.chunked_prefill_size is not None and server_args.enable_mixed_chunk
+                self.chunked_prefill_size is not None and server_args.enable_mixed_chunk
         )
 
         # Init the grammar backend for constrained generation
@@ -287,21 +332,21 @@ class SchedulerCore:
 
         # Init new token estimation
         assert (
-            server_args.schedule_conservativeness >= 0
+                server_args.schedule_conservativeness >= 0
         ), "Invalid schedule_conservativeness"
 
         self.init_new_token_ratio = min(
             global_config.default_init_new_token_ratio
             * server_args.schedule_conservativeness,
             1.0,
-        )
+            )
         self.min_new_token_ratio = min(
             self.init_new_token_ratio
             * global_config.default_min_new_token_ratio_factor,
             1.0,
-        )
+            )
         self.new_token_ratio_decay = (
-                                         self.init_new_token_ratio - self.min_new_token_ratio
+                                             self.init_new_token_ratio - self.min_new_token_ratio
                                      ) / global_config.default_new_token_ratio_decay_steps
         self.new_token_ratio = self.init_new_token_ratio
 
@@ -343,6 +388,21 @@ class SchedulerCore:
                 },
             )
 
+        self._dispatcher = TypeBasedDispatcher([
+            (TokenizedGenerateReqInput, self.handle_generate_request),
+            (TokenizedEmbeddingReqInput, self.handle_embedding_request),
+            (FlushCacheReq, self.flush_cache_wrapped),
+            (AbortReq, self.abort_request),
+            (UpdateWeightFromDiskReqInput, self.update_weights_from_disk),
+            (InitWeightsUpdateGroupReqInput, self.init_weights_update_group),
+            (UpdateWeightsFromDistributedReqInput, self.update_weights_from_distributed),
+            (UpdateWeightsFromTensorReqInput, self.update_weights_from_tensor),
+            (GetWeightsByNameReqInput, self.get_weights_by_name),
+            (ProfileReq, self.profile),
+            (OpenSessionReqInput, self.open_session),
+            (CloseSessionReqInput, self.close_session),
+        ])
+
     def watchdog_thread(self):
         """A watch dog thread that will try to kill the server itself if one batch takes too long."""
         self.watchdog_last_forward_ct = 0
@@ -365,7 +425,7 @@ class SchedulerCore:
     def event_loop_normal(self):
         """A normal scheduler loop."""
         while True:
-            self.callback.on_event_loop_iteration()
+            self._process_input_requests(self._recv_requests())
 
             batch = self.get_next_batch_to_run()
 
@@ -390,7 +450,7 @@ class SchedulerCore:
         result_queue = deque()
 
         while True:
-            self.callback.on_event_loop_iteration()
+            self._process_input_requests(self._recv_requests())
 
             batch = self.get_next_batch_to_run()
             self.cur_batch = batch
@@ -423,9 +483,33 @@ class SchedulerCore:
 
             self.last_batch = batch
 
+    def _recv_requests(self) -> List[Req]:
+        """Receive results at tp_rank = 0 and broadcast it to all other TP ranks."""
+        if self.tp_rank == 0 or self.server_args.enable_dp_attention:
+            recv_reqs = []
+
+            while True:
+                try:
+                    recv_req = self._recv_from_tokenizer.recv_pyobj(zmq.NOBLOCK)
+                except zmq.ZMQError:
+                    break
+                recv_reqs.append(recv_req)
+        else:
+            recv_reqs = None
+
+        if self.server_args.tp_size != 1 and not self.server_args.enable_dp_attention:
+            recv_reqs = broadcast_pyobj(recv_reqs, self.tp_rank, self.core.tp_cpu_group)
+        return recv_reqs
+
+    def _process_input_requests(self, recv_reqs: List):
+        for recv_req in recv_reqs:
+            output = self._dispatcher(recv_req)
+            if output is not None:
+                self._send_to_tokenizer.send_pyobj(output)
+
     def handle_generate_or_embedding_request(
-        self,
-        recv_req: Union[TokenizedGenerateReqInput, TokenizedEmbeddingReqInput],
+            self,
+            recv_req: Union[TokenizedGenerateReqInput, TokenizedEmbeddingReqInput],
     ):
         if isinstance(recv_req, TokenizedGenerateReqInput):
             return self.handle_generate_request(recv_req)
@@ -434,14 +518,14 @@ class SchedulerCore:
         raise NotImplementedError
 
     def handle_generate_request(
-        self,
-        recv_req: TokenizedGenerateReqInput,
+            self,
+            recv_req: TokenizedGenerateReqInput,
     ):
         # Create a new request
         if (
-            recv_req.session_params is None
-            or recv_req.session_params.id is None
-            or recv_req.session_params.id not in self.sessions
+                recv_req.session_params is None
+                or recv_req.session_params.id is None
+                or recv_req.session_params.id not in self.sessions
         ):
 
             if recv_req.input_embeds is not None:
@@ -465,8 +549,8 @@ class SchedulerCore:
             req.tokenizer = self.tokenizer
 
             if (
-                recv_req.session_params is not None
-                and recv_req.session_params.id is not None
+                    recv_req.session_params is not None
+                    and recv_req.session_params.id is not None
             ):
                 req.finished_reason = FINISH_ABORT(
                     f"Invalid request: session id {recv_req.session_params.id} does not exist"
@@ -526,14 +610,14 @@ class SchedulerCore:
                 else 1 << 30
             ),
             self.max_req_len - len(req.origin_input_ids) - 1,
-        )
+            )
 
         # Init grammar cache for this request
         add_to_grammar_queue = False
         if (
-            req.sampling_params.json_schema is not None
-            or req.sampling_params.regex is not None
-            or req.sampling_params.ebnf is not None
+                req.sampling_params.json_schema is not None
+                or req.sampling_params.regex is not None
+                or req.sampling_params.ebnf is not None
         ):
             assert self.grammar_backend is not None
             if req.sampling_params.json_schema is not None:
@@ -554,8 +638,8 @@ class SchedulerCore:
             self.waiting_queue.append(req)
 
     def handle_embedding_request(
-        self,
-        recv_req: TokenizedEmbeddingReqInput,
+            self,
+            recv_req: TokenizedEmbeddingReqInput,
     ):
         req = Req(
             recv_req.rid,
@@ -577,15 +661,15 @@ class SchedulerCore:
 
     def log_prefill_stats(self, adder, can_run_list, running_bs, has_being_chunked):
         self.tree_cache_metrics["total"] += (
-                                                adder.log_input_tokens + adder.log_hit_tokens
+                                                    adder.log_input_tokens + adder.log_hit_tokens
                                             ) / 10 ** 9
         self.tree_cache_metrics["hit"] += (adder.log_hit_tokens) / 10 ** 9
         tree_cache_hit_rate = (
-            self.tree_cache_metrics["hit"] / self.tree_cache_metrics["total"]
+                self.tree_cache_metrics["hit"] / self.tree_cache_metrics["total"]
         )
 
         num_used = self.max_total_num_tokens - (
-            self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
+                self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
         )
 
         logger.info(
@@ -609,10 +693,10 @@ class SchedulerCore:
 
     def log_decode_stats(self):
         num_used = self.max_total_num_tokens - (
-            self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
+                self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
         )
         gen_throughput = self.num_generated_tokens / (
-            time.time() - self.last_decode_stats_tic
+                time.time() - self.last_decode_stats_tic
         )
         self.num_generated_tokens = 0
         self.last_decode_stats_tic = time.time()
@@ -636,7 +720,7 @@ class SchedulerCore:
 
     def check_memory(self):
         available_size = (
-            self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
+                self.token_to_kv_pool.available_size() + self.tree_cache.evictable_size()
         )
         if available_size != self.max_total_num_tokens:
             msg = (
@@ -692,7 +776,7 @@ class SchedulerCore:
 
         # Handle the cases where prefill is not allowed
         if (
-            self.batch_is_full or len(self.waiting_queue) == 0
+                self.batch_is_full or len(self.waiting_queue) == 0
         ) and self.being_chunked_req is None:
             return None
 
@@ -713,7 +797,7 @@ class SchedulerCore:
             self.max_prefill_tokens,
             self.chunked_prefill_size,
             running_bs if self.is_mixed_chunk else 0,
-        )
+            )
 
         has_being_chunked = self.being_chunked_req is not None
         if has_being_chunked:
@@ -730,13 +814,13 @@ class SchedulerCore:
         # Get requests from the waiting queue to a new prefill batch
         for req in self.waiting_queue:
             if (
-                self.lora_paths
-                and len(
+                    self.lora_paths
+                    and len(
                 lora_set
                 | set([req.lora_path for req in adder.can_run_list])
                 | set([req.lora_path])
             )
-                > self.max_loras_per_batch
+                    > self.max_loras_per_batch
             ):
                 self.batch_is_full = True
                 break
@@ -787,9 +871,9 @@ class SchedulerCore:
 
         # Mixed-style chunked prefill
         if (
-            self.is_mixed_chunk
-            and self.running_batch is not None
-            and not (new_batch.return_logprob or self.running_batch.return_logprob)
+                self.is_mixed_chunk
+                and self.running_batch is not None
+                and not (new_batch.return_logprob or self.running_batch.return_logprob)
         ):
             # TODO (lianmin): support return_logprob + mixed chunked prefill
             self.running_batch.filter_batch()
@@ -816,7 +900,7 @@ class SchedulerCore:
 
         # Check if decode out of memory
         if not batch.check_decode_mem(self.decode_mem_cache_buf_multiplier) or (
-            test_retract and batch.batch_size() > 10
+                test_retract and batch.batch_size() > 10
         ):
             old_ratio = self.new_token_ratio
 
@@ -835,7 +919,7 @@ class SchedulerCore:
             self.new_token_ratio = max(
                 self.new_token_ratio - self.new_token_ratio_decay,
                 self.min_new_token_ratio,
-            )
+                )
 
         # Check for jump-forward
         if not self.disable_jump_forward:
@@ -1049,18 +1133,18 @@ class SchedulerCore:
 
         self.forward_ct_decode = (self.forward_ct_decode + 1) % (1 << 30)
         if (
-            self.tp_rank == 0
-            and self.forward_ct_decode % self.server_args.decode_log_interval == 0
+                self.tp_rank == 0
+                and self.forward_ct_decode % self.server_args.decode_log_interval == 0
         ):
             self.log_decode_stats()
 
     def add_logprob_return_values(
-        self,
-        i: int,
-        req: Req,
-        pt: int,
-        next_token_ids: List[int],
-        output: LogitsProcessorOutput,
+            self,
+            i: int,
+            req: Req,
+            pt: int,
+            next_token_ids: List[int],
+            output: LogitsProcessorOutput,
     ):
         """Attach logprobs to the return values."""
         req.output_token_logprobs_val.append(output.next_token_logprobs[i])
@@ -1091,7 +1175,7 @@ class SchedulerCore:
             ]
 
             if (
-                req.logprob_start_len == 0
+                    req.logprob_start_len == 0
             ):  # The first token does not have logprob, pad it.
                 input_token_logprobs_val = [None] + input_token_logprobs_val
                 input_token_logprobs_idx = [req.fill_ids[0]] + input_token_logprobs_idx
@@ -1140,7 +1224,7 @@ class SchedulerCore:
         return num_input_logprobs
 
     def stream_output(
-        self, reqs: List[Req], return_logprob: bool, skip_req: Optional[Req] = None
+            self, reqs: List[Req], return_logprob: bool, skip_req: Optional[Req] = None
     ):
         """Stream the output to detokenizer."""
         rids = []
@@ -1186,11 +1270,11 @@ class SchedulerCore:
 
                 # TODO(lianmin): revisit this for overlap + retract + stream
                 if (
-                    req.finished()
-                    # If stream, follow the given stream_interval
-                    or (req.stream and len(req.output_ids) % self.stream_interval == 0)
-                    # If not stream, we still want to output some tokens to get the benefit of incremental decoding.
-                    or (not req.stream and len(req.output_ids) % 50 == 0)
+                        req.finished()
+                        # If stream, follow the given stream_interval
+                        or (req.stream and len(req.output_ids) % self.stream_interval == 0)
+                        # If not stream, we still want to output some tokens to get the benefit of incremental decoding.
+                        or (not req.stream and len(req.output_ids) % 50 == 0)
                 ):
                     if self.draft_worker and req.finished():
                         self.draft_worker.finish_request(req)
@@ -1235,7 +1319,7 @@ class SchedulerCore:
 
             # Send to detokenizer
             if rids:
-                self.callback.on_output(
+                self._send_to_detokenizer.send_pyobj(
                     BatchTokenIDOut(
                         rids,
                         finished_reasons,
@@ -1271,7 +1355,7 @@ class SchedulerCore:
                     finished_reasons.append(req.finished_reason.to_json())
                     embeddings.append(req.embedding)
                     prompt_tokens.append(len(req.origin_input_ids))
-            self.callback.on_output(
+            self._send_to_detokenizer.send_pyobj(
                 BatchEmbeddingOut(rids, finished_reasons, embeddings, prompt_tokens)
             )
 
@@ -1361,7 +1445,7 @@ class SchedulerCore:
     def flush_cache(self):
         """Flush the memory pool and cache."""
         if len(self.waiting_queue) == 0 and (
-            self.running_batch is None or len(self.running_batch.reqs) == 0
+                self.running_batch is None or len(self.running_batch.reqs) == 0
         ):
             self.tree_cache.reset()
             self.tree_cache_metrics = {"total": 0, "hit": 0}
@@ -1418,7 +1502,7 @@ class SchedulerCore:
         return InitWeightsUpdateGroupReqOutput(success, message)
 
     def update_weights_from_distributed(
-        self, recv_req: UpdateWeightsFromDistributedReqInput
+            self, recv_req: UpdateWeightsFromDistributedReqInput
     ):
         """Update the online model parameter."""
         success, message = self.tp_worker.update_weights_from_distributed(recv_req)
@@ -1492,3 +1576,50 @@ class SchedulerCore:
 class SchedulerCoreCallback:
     on_output: Callable
     on_event_loop_iteration: Callable
+
+
+def run_scheduler_process(
+        server_args: ServerArgs,
+        port_args: PortArgs,
+        gpu_id: int,
+        tp_rank: int,
+        dp_rank: Optional[int],
+        pipe_writer,
+):
+    setproctitle.setproctitle("sglang::scheduler")
+
+    # [For Router] if env var "SGLANG_DP_RANK" exist, set dp_rank to the value of the env var
+    if dp_rank is None and "SGLANG_DP_RANK" in os.environ:
+        dp_rank = int(os.environ["SGLANG_DP_RANK"])
+
+    # Configue the logger
+    if dp_rank is None:
+        configure_logger(server_args, prefix=f" TP{tp_rank}")
+    else:
+        configure_logger(server_args, prefix=f" DP{dp_rank} TP{tp_rank}")
+    suppress_other_loggers()
+
+    # Set cpu affinity to this gpu process
+    if get_bool_env_var("SGLANG_SET_CPU_AFFINITY"):
+        set_gpu_proc_affinity(server_args.tp_size, server_args.nnodes, gpu_id)
+
+    parent_process = psutil.Process().parent()
+
+    # Create a scheduler and run the event loop
+    try:
+        scheduler = SchedulerCore(
+            server_args=server_args, nccl_port=port_args.nccl_port,
+            gpu_id=gpu_id, tp_rank=tp_rank, dp_rank=dp_rank,
+        )
+
+        pipe_writer.send(
+            {"status": "ready", "max_total_num_tokens": scheduler.max_total_num_tokens}
+        )
+        if scheduler.enable_overlap:
+            scheduler.event_loop_overlap()
+        else:
+            scheduler.event_loop_normal()
+    except Exception:
+        traceback = get_exception_traceback()
+        logger.error(f"Scheduler hit an exception: {traceback}")
+        parent_process.send_signal(signal.SIGQUIT)
